requirements-dev.txt
# Development and testing dependencies
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-flask>=1.2.0

# Linting
flake8>=6.0.0
black>=23.0.0

# Security
bandit>=1.7.0
safety>=2.0.0
